Split build tool to separat files (#678)
