+++ tools/buildTools/dts.js
@@ -1,24 +1,37 @@
 'use strict';
-var fs = require('fs');
-var path = require('path');
-var namePlaceholder = '__NAME__';
-var regExportFrom = /export([^;]+)from\s+'([^']+)';/gm;
-var regImportFrom = /import[^;]+from\s+'([^']+)';/gm;
-var singleLineComment = /\/\/[^\n]*\n/g;
-var multiLineComment = /(^\/\*(\*(?!\/)|[^*])*\*\/\s*)/m;
+
+const fs = require('fs');
+const path = require('path');
+const glob = require('glob');
+const mkdirp = require('mkdirp');
+const {
+    rootPath,
+    distPath,
+    roosterJsDistPath,
+    mainPackageJson,
+    nodeModulesPath,
+    runNode,
+    err,
+} = require('./common');
+
+const namePlaceholder = '__NAME__';
+const regExportFrom = /export([^;]+)from\s+'([^']+)';/gm;
+const regImportFrom = /import[^;]+from\s+'([^']+)';/gm;
+const singleLineComment = /\/\/[^\n]*\n/g;
+const multiLineComment = /(^\/\*(\*(?!\/)|[^*])*\*\/\s*)/m;
 
 // 1. [export ][default |declare ](class|interface) <NAME>[ extends| implements <BASE_CLASS>] {...}
-var regClassInterface = /(\/\*(\*(?!\/)|[^*])*\*\/\s*)?(export\s+)?(default\s+|declare\s+)?(interface|class)\s+([a-zA-Z0-9_]+(\s*<[^>]+>)?)((\s+extends|\s+implements)(\s[0-9a-zA-Z_\.\s,]+(\s*<[^{]+>)?))?\s*{/g;
+const regClassInterface = /(\/\*(\*(?!\/)|[^*])*\*\/\s*)?(export\s+)?(default\s+|declare\s+)?(interface|class)\s+([a-zA-Z0-9_]+(\s*<[^>]+>)?)((\s+extends|\s+implements)(\s[0-9a-zA-Z_\.\s,]+(\s*<[^{]+>)?))?\s*{/g;
 // 2. [export ][default |declare ]function <NAME>(...)[: <TYPE>];
-var regFunction = /(\/\*(\*(?!\/)|[^*])*\*\/\s*)?(export\s+)?(default\s+|declare\s+)?function\s+([a-zA-Z0-9_]+(\s*<(?:[^>]|=>)+>)?)\s*(\([^;]+;)/g;
+const regFunction = /(\/\*(\*(?!\/)|[^*])*\*\/\s*)?(export\s+)?(default\s+|declare\s+)?function\s+([a-zA-Z0-9_]+(\s*<(?:[^>]|=>)+>)?)\s*(\([^;]+;)/g;
 // 3. [export ][default |declare ]const enum <NAME> {...}
-var regEnum = /(\/\*(\*(?!\/)|[^*])*\*\/\s*)?(export\s+)?(default\s+|declare\s+)?(const\s+)?enum\s+([a-zA-Z0-9_<>]+)\s*{/g;
+const regEnum = /(\/\*(\*(?!\/)|[^*])*\*\/\s*)?(export\s+)?(default\s+|declare\s+)?(const\s+)?enum\s+([a-zA-Z0-9_<>]+)\s*{/g;
 // 4. [export ][default |declare ]type <NAME> = ...;
-var regType = /(\/\*(\*(?!\/)|[^*])*\*\/\s*)?(export\s+)?(default\s+|declare\s+)?type\s+([0-9a-zA-Z_]+(\s*<[^>]+>)?)\s*=\s*/g;
+const regType = /(\/\*(\*(?!\/)|[^*])*\*\/\s*)?(export\s+)?(default\s+|declare\s+)?type\s+([0-9a-zA-Z_]+(\s*<[^>]+>)?)\s*=\s*/g;
 // 5. [export ][default |declare ]const <NAME>: ...;
-var regConst = /(\/\*(\*(?!\/)|[^*])*\*\/\s*)?(export\s+)?(default\s+|declare\s+)?const\s+([0-9a-zA-Z_<>]+)\s*:\s*/g;
+const regConst = /(\/\*(\*(?!\/)|[^*])*\*\/\s*)?(export\s+)?(default\s+|declare\s+)?const\s+([0-9a-zA-Z_<>]+)\s*:\s*/g;
 // 6. export[ default] <NAME>|{NAMES};
-var regExport = /(\/\*(\*(?!\/)|[^*])*\*\/\s*)?(export\s+)(default\s+([0-9a-zA-Z_]+)\s*,?)?(\s*{([^}]+)})?\s*;/g;
+const regExport = /(\/\*(\*(?!\/)|[^*])*\*\/\s*)?(export\s+)(default\s+([0-9a-zA-Z_]+)\s*,?)?(\s*{([^}]+)})?\s*;/g;
 
 function enqueue(queue, filename, exports) {
     var existingItem = queue.find(v => v.filename == filename);
@@ -252,9 +265,7 @@ function publicElement(element) {
 }
 
 function output(targetDir, library, isAmd, queue) {
-    var version = JSON.stringify(
-        require(path.join(__dirname, '..', 'package.json')).version
-    ).replace(/"/g, '');
+    var version = JSON.stringify(mainPackageJson.version).replace(/"/g, '');
     var content = '';
     content += `// Type definitions for roosterjs (Version ${version})\r\n`;
     content += '// Generated by dts tool from roosterjs\r\n';
@@ -334,13 +345,7 @@ function output(targetDir, library, isAmd, queue) {
     return filename;
 }
 
-function err(message) {
-    let ex = new Error('\n' + message);
-    console.error(ex.message);
-    throw ex;
-}
-
-module.exports.prepareDts = (rootPath, baseDir, root, additionalFiles) => {
+function createQueue(rootPath, baseDir, root, additionalFiles) {
     var queue = [];
     var i = 0;
 
@@ -357,6 +362,34 @@ module.exports.prepareDts = (rootPath, baseDir, root, additionalFiles) => {
     }
 
     return queue;
-};
+}
 
-module.exports.output = output;
+function dts(isAmd) {
+    mkdirp.sync(roosterJsDistPath);
+
+    const tsFiles = glob
+        .sync(path.relative(rootPath, path.join(distPath, '**', 'lib', '**', '*.d.ts')), {
+            nocase: true,
+        })
+        .map(x => path.relative(distPath, x));
+    const dtsQueue = createQueue(rootPath, distPath, 'roosterjs/lib/index.d.ts', tsFiles);
+    const filename = output(roosterJsDistPath, 'roosterjs', isAmd, dtsQueue);
+
+    if (!isAmd) {
+        const typescriptPath = path.join(nodeModulesPath, 'typescript/lib/tsc.js');
+        runNode(typescriptPath + ' ' + filename + ' --noEmit', rootPath);
+    }
+}
+
+module.exports = {
+    dtsCommonJs: {
+        message: `Generating type definition file for CommonJs}...`,
+        callback: () => dts(false /*isAmd*/),
+        enabled: options => options.dts,
+    },
+    dtsAmd: {
+        message: `Generating type definition file for AMD}...`,
+        callback: () => dts(true /*isAmd*/),
+        enabled: options => options.dts,
+    },
+};
